Don't schedule a `requestAnimationFrame` if there is nothing to animate
With this change, a `requestAnimationFrame` is not scheduled unless there is active confetti to animate.

This is because when idle, the browser will continue to do work, even though there is no work to do.

However, adding confentti will restart the loop if it is not currently running. If it is, no additional frames are scheduled.