Return a promise when an animation batch completes animating
To enable customized behaviour when a confetti burst has completed it's animation, we need some sort of notification.

With this change, when confetti is added a `Promise` is returned that will complete when all member shapes in that batch have been removed.

Also updates exposed types, and the sample page.