Address PR feedback
- Rename `members` to `shapes`
- Rename temporary variable in `addConfetti` to more clearly denote what the item is
- Removed lazy instantiation of the promise in `ConfettiBatch` (See note below)

Original desire was to 'only create the promise if someone actually uses it'. Except that is not really how its used -- poor choice on my part with the orignal code.

However, moving it to always create doesn't remove the ugly wart of optional chaining. This is 'cause the flow from the promise constructor is passed a function, and flow analsysis doesn't *know* the field is assigned to, so demands optionality.

One option would be to assign a function to it by default, that is then overwritten, and all would be good. However that feels noisy given the allocation of the function that is never used.