Address confetti not hitting peak-x with stomped `currentTime`
The movement derived from the delta time was being compressed by a shorter-than-previous `lastUpdated` time -- the cost of performing the calculation is a factor in the animation itself.

This changes the capturing of `currentTime` to use the one captured before the processing of the shapes. But if this was the first iteration of the loop (aka queing it), we'll just capture it and start the world again.