Correctly position confetti source when confetti added immediately after instantiation
With the previous usage of `canvas.width|.height`, if confetti was added immediately after the canvas was created, the position of the confetti source was incorrect. This is because no layout pass had completed, and the defaults of 300x150 for the height/width of the canvas were applied (see: https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement).

Switching to `getBoundingClientRect` forces a layout pass in the browser, and returns the correct sizes immediately on creation.