Don't schedule a `requestAnimationFrame` if there is nothing to animate (#36)
* Don't schedule a `requestAnimationFrame` if there is nothing to animate

With this change, a `requestAnimationFrame` is not scheduled unless there is active confetti to animate.

This is because when idle, the browser will continue to do work, even though there is no work to do.

However, adding confentti will restart the loop if it is not currently running. If it is, no additional frames are scheduled.

* Address confetti not hitting peak-x with stomped `currentTime`

The movement derived from the delta time was being compressed by a shorter-than-previous `lastUpdated` time -- the cost of performing the calculation is a factor in the animation itself.

This changes the capturing of `currentTime` to use the one captured before the processing of the shapes. But if this was the first iteration of the loop (aka queing it), we'll just capture it and start the world again.