Return a promise when an confetti batch completes animating (#37)
* Return a promise when an animation batch completes animating

To enable customized behaviour when a confetti burst has completed it's animation, we need some sort of notification.

With this change, when confetti is added a `Promise` is returned that will complete when all member shapes in that batch have been removed.

Also updates exposed types, and the sample page.

* Address PR feedback

- Rename `members` to `shapes`
- Rename temporary variable in `addConfetti` to more clearly denote what the item is
- Removed lazy instantiation of the promise in `ConfettiBatch` (See note below)

Original desire was to 'only create the promise if someone actually uses it'. Except that is not really how its used -- poor choice on my part with the orignal code.

However, moving it to always create doesn't remove the ugly wart of optional chaining. This is 'cause the flow from the promise constructor is passed a function, and flow analsysis doesn't *know* the field is assigned to, so demands optionality.

One option would be to assign a function to it by default, that is then overwritten, and all would be good. However that feels noisy given the allocation of the function that is never used.

* Revert unintentional change to demo page

* Refactor processing of batches/shapes

Reduce the number of passes through shapes/batches for rendering & cleaning up.

- Moved primary processing of *shapes* into the batch class itself
- Change loop to iterate over the batches, not shapes. Each batch handles the shape work
- Changed the work to `filter` not `forEach`: this minimizes the number of passes over the list, trading off against allocations of an array every time. Now there is `batch * shapes` operations every time with no `2 * shapes` every 100 iterations.