refactor angles related logic in ConfettiShape
Get rid of "direction" property that was passed to ConfettiShape by passing initialFlightAngle and rotationAngle props.
This would allow us next to init confetti not just from the sides of the screen, but also from any given position, for example on mouse click.

Also, it seems that the cos and sin were mixed up in the code, sin now updates X position and cos updates Y position. Assuming that the 0 angle is at the top