feat: implement addConfettiAtPosition first prototype
Implement `__DO_NOT_USE_THIS_IS_UNDER_DEVELOPMENT__addConfettiAtPosition` method in `JSConfetti` that is used for dispatching confetti from specific position on the screen (for example on mouse click).

This is an early prototype and I might change a few things, so I've marked the method and params with `__DO_NOT_USE*` prefixes.

In `ConfettiShape` I've added `isVisible` flag to control confetti visibility with offset position: when confetti is dispatched, it has random shifted position from the initial position, it is done for aesthetic purposes to create an effect that confetti takes some random time to be dispatched, we need to hide confetti until it has reached its initial position. However, I hide it only when with a flag, to keep the behaviour of "From sides of the screen" as is, confetti are still half-way visible there when they didn't reach the `dispatchPosition` and it looks good